Split up some server logic
This is a first pass at splitting the server logic into multiple layers.
Current thinking is that the "frontend" layer is responsible for the
actual transport protocol (parsing requests, formatting responses, and
so on) and client-type-specific stuff (since that's really just part of
"formatting responses"). In future that could be split further, so
there's a transport layer and a client-specific-formatting layer or
something like that.

The backend does the rest.