Allow URL to be set
Do this with an extra button that causes a dialog to show.

It's kind of ugly though, so we might want to change this in the future.
One idea is to show the dialog on startup (before location, since it's more
important than that, although that needs to be fixed separately), and
just have a little menu button for changing it.