Bump target API version, remove binding, and request background location
The API version update is to satisfy Play policies, and just requires
fixing some insets.

The binding removal was an attempt to make foreground service sufficient
for my location purposes, but it didn't work. I kept it because it seems
a bit cleaner though.

I finally relented and added background location permission, which is
going to require a proper permission explanation too (again to satisfy
Play policies). In the process I realised that notification permission
is optional, so I made that change too.